+++ ChangeLog.md
@@ -4,6 +4,7 @@
 - Added active flag to players/goalies
 - Clear rookie flag on new (regular) season
 - Save a copy of the database on new season
+- Implemented game standings editing
 
 ## 0.10.0
 - Don't show player number zero in reports