+++ ChangeLog.md
@@ -5,6 +5,7 @@
 - Clear rookie flag on new (regular) season
 - Save a copy of the database on new season
 - Implemented game standings editing
+- Added title screen
 
 ## 0.10.0
 - Don't show player number zero in reports